tests/test_main.py
# tests/test_main.py

import unittest
from unittest.mock import patch, MagicMock
from main import process_text

class TestMain(unittest.TestCase):
    @patch('main.detect_text')
    @patch('main.analyze_signature')
    @patch('main.analyze_style')
    @patch('main.morph_style')
    @patch('main.generate_audit_report')
    @patch('main.visualize_analysis')
    def test_process_text(self, mock_visualize, mock_audit, mock_morph, mock_style, mock_signature, mock_detect):
        # Setup mock return values
        mock_detect.return_value = {'score': 0.8, 'summary': 'Likely human-authored'}
        mock_signature.return_value = {
            'summary': {'language_complexity': 'Medium', 'vocabulary_diversity': 'High'},
            'avg_word_length': 5.0,
            'unique_word_ratio': 0.7
        }
        mock_style.return_value = {'summary': 'Varied', 'word_count_variance': 2.5}
        mock_morph.return_value = "This is a morphed sentence."
        mock_audit.return_value = {
            "timestamp": "2023-06-01T12:00:00",
            "text_stats": {
                "word_count": 5,
                "character_count": 23,
                "average_word_length": 4.6
            },
            "metadata": {"source": "user_input"},
            "audit_summary": "Audit completed for text with 5 words and 23 characters."
        }
        mock_visualize.return_value = "base64_encoded_image_string"

        test_text = "This is a test sentence."
        
        # Capture printed output
        with patch('builtins.print') as mock_print:
            process_text(test_text)

        # Assert that all mock functions were called
        mock_detect.assert_called_once_with(test_text)
        mock_signature.assert_called_once_with(test_text)
        mock_style.assert_called_once_with(test_text)
        mock_morph.assert_called_once_with(test_text, "formal")
        mock_audit.assert_called_once_with(test_text, {"source": "user_input"})
        mock_visualize.assert_called_once()

        # Assert that print was called (output was generated)
        self.assertTrue(mock_print.called)

    def test_process_text_error_handling(self):
        with patch('main.detect_text', side_effect=Exception("Test exception")):
            with self.assertRaises(SystemExit):
                process_text("Test text")

if __name__ == '__main__':
    unittest.main()import unittest
from unittest.mock import patch
from main import process_text

class TestMain(unittest.TestCase):
    @patch('main.detect_text')
    @patch('main.analyze_signature')
    @patch('main.analyze_style')
    @patch('main.morph_style')
    @patch('main.generate_audit_report')
    def test_process_text(self, mock_audit, mock_morph, mock_style, mock_signature, mock_detect):
        # Setup mock return values
        mock_detect.return_value = {'score': 0.8, 'summary': 'Likely human-authored'}
        mock_signature.return_value = {
            'summary': {'language_complexity': 'Medium', 'vocabulary_diversity': 'High'},
            'avg_word_length': 5.0,
            'unique_word_ratio': 0.7
        }
        mock_style.return_value = {'summary': 'Varied', 'word_count_variance': 2.5}
        mock_morph.return_value = "This is a morphed sentence."
        mock_audit.return_value = {"audit_summary": "Placeholder audit report"}

        test_text = "This is a test sentence."
        
        # Capture printed output
        with patch('builtins.print') as mock_print:
            process_text(test_text)

        # Assert that all mock functions were called
        mock_detect.assert_called_once_with(test_text)
        mock_signature.assert_called_once_with(test_text)
        mock_style.assert_called_once_with(test_text)
        mock_morph.assert_called_once_with(test_text, "formal")
        mock_audit.assert_called_once_with(test_text, {"source": "user_input"})

        # Assert that print was called (output was generated)
        self.assertTrue(mock_print.called)

    def test_process_text_error_handling(self):
        with patch('main.detect_text', side_effect=Exception("Test exception")):
            with self.assertRaises(SystemExit):
                process_text("Test text")

if __name__ == '__main__':
    unittest.main()import unittest
from main import process_text

class TestMain(unittest.TestCase):
    def test_process_text(self):
        # This is a very basic test and should be expanded
        test_text = "This is a test sentence."
        # We're just checking if the function runs without errors
        try:
            process_text(test_text)
        except Exception as e:
            self.fail(f"process_text raised {type(e).__name__} unexpectedly!")

if __name__ == '__main__':
    unittest.main()